Segregate bulk tests
