+++ src/test/java/org/folio/rest/impl/accountactionchecks/AccountsActionChecksAPITestsBase.java
@@ -1,0 +1,184 @@
+package org.folio.rest.impl.accountactionchecks;
+
+import static io.restassured.http.ContentType.JSON;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.nullValue;
+
+import java.util.Arrays;
+import java.util.stream.Stream;
+
+import org.apache.http.HttpStatus;
+import org.folio.rest.domain.FeeFineStatus;
+import org.folio.rest.jaxrs.model.Account;
+import org.folio.rest.jaxrs.model.BulkCheckActionRequest;
+import org.folio.rest.jaxrs.model.CheckActionRequest;
+import org.folio.rest.utils.ResourceClient;
+import org.folio.test.support.ApiTests;
+import org.folio.test.support.EntityBuilder;
+import org.hamcrest.CoreMatchers;
+
+public class AccountsActionChecksAPITestsBase extends ApiTests {
+
+  protected Account firstAccount;
+  protected Account secondAccount;
+
+  public static final String ACCOUNTS_TABLE = "accounts";
+  public static final String ERROR_MESSAGE_ALREADY_CLOSED = "Fee/fine is already closed";
+
+  private static final double ACCOUNT_INITIAL_AMOUNT = 9.00;
+  private static final double ACCOUNT_REMAINING_AMOUNT = 4.55;
+  private static final double REQUESTED_AMOUNT = 1.23;
+  private static final String REQUESTED_AMOUNT_STRING = String.valueOf(REQUESTED_AMOUNT);
+
+  protected void actionShouldNotBeAllowed(boolean bulk,
+                                          ResourceClient accountsActionCheckClient, String amount, String errorMessage) {
+
+    accountsActionCheckClient.attemptCreate(createRequest(bulk, amount))
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body(containsString(errorMessage))
+      .body("allowed", is(false))
+      .body("accountIds",
+        bulk ? CoreMatchers.is(Arrays.asList(firstAccount.getId(), secondAccount.getId()))
+          : nullValue())
+      .body("amount", is(amount));
+  }
+
+  protected void actionShouldBeAllowed(boolean bulk, ResourceClient actionCheckClient,
+                                       String remaining) {
+
+    actionCheckClient.attemptCreate(createRequest(bulk,
+      REQUESTED_AMOUNT_STRING))
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("allowed", is(true))
+      .body("accountIds", bulk ? is(Arrays.asList(firstAccount.getId(), secondAccount.getId()))
+        : nullValue())
+      .body("amount", is(REQUESTED_AMOUNT_STRING))
+      .body("remainingAmount", is(remaining));
+  }
+
+  protected void actionCheckAmountShouldNotBeAllowedWithExceededAmount(boolean bulk,
+                                                                       ResourceClient actionCheckClient) {
+
+    String expectedErrorMessage = "Requested amount exceeds remaining amount";
+    String amount = String.valueOf(bulk ? ACCOUNT_REMAINING_AMOUNT * 2 + 1 : REQUESTED_AMOUNT
+      + 10);
+
+    baseActionCheckAmountShouldNotBeAllowedWithExceededAmount(actionCheckClient,
+      expectedErrorMessage, amount);
+  }
+
+  protected void actionCheckRefundAmountShouldNotBeAllowedWithExceededAmount(
+    ResourceClient actionCheckClient) {
+
+    String expectedErrorMessage =
+      "Refund amount must be greater than zero and less than or equal to Selected amount";
+
+    baseActionCheckAmountShouldNotBeAllowedWithExceededAmount(
+      actionCheckClient, expectedErrorMessage, String.valueOf(
+        REQUESTED_AMOUNT + 10));
+  }
+
+  void baseActionCheckAmountShouldNotBeAllowedWithExceededAmount(
+    ResourceClient actionCheckClient, String expectedErrorMessage, String amount) {
+
+    CheckActionRequest accountCheckRequest = new CheckActionRequest();
+    accountCheckRequest.withAmount(amount);
+
+    actionCheckClient.attemptCreate(accountCheckRequest)
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body("errorMessage", is(expectedErrorMessage))
+      .body("allowed", is(false))
+      .body("amount", is(accountCheckRequest.getAmount()));
+  }
+
+  protected void actionCheckShouldNotFailForNonExistentAccount(boolean bulk,
+                                                               ResourceClient actionCheckClient) {
+
+    actionCheckClient.attemptCreate(createRequest(bulk,
+      REQUESTED_AMOUNT_STRING))
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND);
+  }
+
+  protected void actionCheckAmountShouldNotBeAllowedForClosedAccount(boolean bulk,
+                                                                     ResourceClient client) {
+
+    closeAllAccounts();
+
+    client.attemptCreate(createRequest(bulk, REQUESTED_AMOUNT_STRING))
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body("allowed", is(false))
+      .body("accountIds", bulk ? is(Arrays.asList(firstAccount.getId(), secondAccount.getId()))
+        : nullValue())
+      .body("amount", is(REQUESTED_AMOUNT_STRING))
+      .body("errorMessage", is(ERROR_MESSAGE_ALREADY_CLOSED));
+
+    removeAllFromTable(ACCOUNTS_TABLE);
+  }
+
+  private void closeAllAccounts() {
+    Stream.of(firstAccount, secondAccount)
+      .forEach(account -> {
+        account.setRemaining(0.00);
+        account.getStatus().setName(FeeFineStatus.CLOSED.getValue());
+        accountsClient.update(account.getId(), account);
+      });
+  }
+
+
+  protected void successfulActionCheckHandlesLongDecimalsCorrectly(ResourceClient client) {
+    firstAccount.setRemaining(1.235987654321); // will be rounded to 1.24
+    accountsClient.update(firstAccount.getId(), firstAccount);
+
+    client.attemptCreate(new CheckActionRequest().withAmount("1.004987654321")) // rounded to 1.00
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("allowed", is(true))
+      .body("amount", is("1.00"))
+      .body("remainingAmount", is("0.24")); // 1.24 - 1.00
+
+    removeAllFromTable(ACCOUNTS_TABLE);
+  }
+
+  protected void failedActionCheckReturnsInitialRequestedAmount(ResourceClient client) {
+    firstAccount.setRemaining(0.99);
+    accountsClient.update(firstAccount.getId(), firstAccount);
+
+    String requestedAmount = "1.004123456789"; // rounded to 1.00 when compared to account balance
+
+    client.attemptCreate(new CheckActionRequest().withAmount(requestedAmount))
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY)
+      .body("allowed", is(false))
+      .body("amount", is(requestedAmount))
+      .body("errorMessage", is("Requested amount exceeds remaining amount"));
+
+    removeAllFromTable(ACCOUNTS_TABLE);
+  }
+
+  protected Account createAccount() {
+    Account accountToPost = EntityBuilder
+      .buildAccount(ACCOUNT_INITIAL_AMOUNT,
+        ACCOUNT_REMAINING_AMOUNT);
+    accountsClient.create(accountToPost)
+      .then()
+      .statusCode(HttpStatus.SC_CREATED)
+      .contentType(JSON);
+    return accountToPost;
+  }
+
+  protected Object createRequest(boolean bulk, String amount) {
+    if (bulk) {
+      return new BulkCheckActionRequest()
+        .withAccountIds(Arrays.asList(firstAccount.getId(), secondAccount.getId()))
+        .withAmount(amount);
+    } else {
+      return new CheckActionRequest().withAmount(amount);
+    }
+  }
+}
